refactor: move ReleaseAuthorizationHandlersTestUtil into Utils namespace
